68 - Modularizando isLate e getStageDate
